+++ src/llm-providers/remote/interfaces/RemoteLlmExecutionToolsOptions.ts
@@ -1,7 +1,5 @@
 import type { CommonToolsOptions } from '../../../execution/CommonToolsOptions';
-import type { string_base_url } from '../../../types/typeAliases';
-import type { string_uri } from '../../../types/typeAliases';
-import type { string_user_id } from '../../../types/typeAliases';
+import type { string_base_url, string_uri, string_user_id } from '../../../types/typeAliases';
 import type { LlmToolsConfiguration } from '../../_common/register/LlmToolsConfiguration';
 import type { CollectionRemoteServerClientOptions } from './RemoteServerOptions';
 
@@ -45,7 +43,7 @@ export type RemoteLlmExecutionToolsOptions<TCustomOptions> = CommonToolsOptions 
                * Identifier of the end user
                *
                * Note: This is passed to the certain model providers to identify misuse
-               * Note: In anonymous mode, there is no need to identify yourself, nor does it change the actual configuration of LLM Tools (unlike in collection mode).
+               * Note: In anonymous mode, there is no need to identify yourself, nor does it change the actual configuration of LLM Tools (unlike in application mode).
                */
               readonly userId: string_user_id | null;
           }